% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat_simu.R
\name{lsrat_simu.conti}
\alias{lsrat_simu.conti}
\title{Simulation for LSRAT and RSMMAT test}
\usage{
lsrat_simu.conti(
  n.sample = 5000,
  n.time = 7,
  par = list(),
  time_cov = TRUE,
  snp.count = 100,
  intercept = TRUE,
  power = FALSE,
  seed = NULL
)
}
\arguments{
\item{n.sample}{Numeric, sample size, number of individuals}

\item{n.time}{Numeric, number of measurements for each individual}

\item{par}{List, the parameters for the phenotype traits, including covaraites and individual specific time dependent random effects}

\item{time_cov}{Logical variable, indicating whether time effect is included in phenotypic traits}

\item{snp.count}{Numeric, number of SNPs in each variant set}

\item{intercept}{Logical variable, indicating whether intercept is used in phenotypic traits}

\item{power}{Logical variable, indicating whether the phenotype generated under the null model (type I error) or under the alternative model (power test)}
}
\value{
A list object is returned to be used as object for association test
}
\description{
This function use pre-defined parameters to make the simulation data with longitudinal continuous traits for the LSRAT test (including type I and power test)
}
