% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat_SSD.R
\name{lsrat.SSD.All}
\alias{lsrat.SSD.All}
\title{LSRAT test or RSMMAT test using SSD format files}
\usage{
lsrat.SSD.All(SSD.INFO, lsrat_est.obj, ...)
}
\arguments{
\item{SSD.INFO}{SSD format information file, output of function "Open_SSD". The genome wide scan are run set by set.}

\item{lsrat_est.obj}{ouput from lsrat_est() or rsmmat_est()}

\item{...}{Other options of the LSRAT or RSMMAT test. Deﬁned same as in function "lsrat_test()" or "rsmmat_est()"}
}
\value{
reults of the LSRAT test or RSMMAT test. First column contains batchID, second column contains SNP ID, third column concains prospective P-value and forth column contains retrospective P-value
}
\description{
LSRAT test or RSMMAT test using SSD format files
}
