% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat_estimate.R
\name{lsrat_est}
\alias{lsrat_est}
\title{GEE NULL model estimation}
\usage{
lsrat_est(y.long, time, y.cov, timecov = TRUE, corstr = "ar1")
}
\arguments{
\item{y.long}{Long-formatted phenotype vector}

\item{time}{Time covarites matched with phenotype vector}

\item{y.cov}{Covariate matrix denoting the covariate variables measured at each time}

\item{timecov}{Logical variable, indicating whether the time fixed effect is estimated}

\item{corstr}{String, correlation structure for GEE model, optional values are: 'ar1', 'ind', 'mixture'}
}
\value{
This function returns a list object with model parameters and residuals of the NULL GEE model
}
\description{
This function estimate the parameters and residuals for the NULL model for GEE/LSRAT tests
}
