% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat_estimate.R
\name{rsmmat_est}
\alias{rsmmat_est}
\title{GLMM NULL model estimation}
\usage{
rsmmat_est(y.long, time, y.cov, timecov = TRUE)
}
\arguments{
\item{y.long}{Long-formatted phenotype vector}

\item{time}{Time covarites matched with phenotype vector}

\item{y.cov}{Covariate matrix denoting the covariate variables measured at each time}

\item{timecov}{Logical variable, indicating whether the time fixed effect is estimated}
}
\value{
This function returns a list object with model parameters and residuals of the NULL GLMM model
}
\description{
This function estimate the parameters and residuals for the NULL model in SMMAT/RSMMAT test
}
